refactor givings section
