+++ src/pages/promoters/FundPage/GivingsSection/AllGivingsSection/index.tsx
@@ -1,0 +1,91 @@
+import { useTranslation } from "react-i18next";
+import Carousel from "components/moleculars/sliders/Carousel";
+import CardDoubleTextDividerButton from "components/moleculars/cards/CardDoubleTextDividerButton";
+import useBreakpoint from "hooks/useBreakpoint";
+import { logError } from "services/crashReport";
+import { formatFromWei } from "lib/web3Helpers/etherFormatters";
+import { formatDate } from "lib/web3Helpers/timeStampFormatters";
+import { useEffect, useState, useCallback } from "react";
+import { logEvent } from "services/analytics";
+import useNavigation from "hooks/useNavigation";
+import usePromoterDonations from "hooks/apiTheGraphHooks/usePromoterDonations";
+import { useWalletContext } from "contexts/walletContext";
+import { useNetwork } from "hooks/useNetwork";
+import RightArrowBlack from "../assets/right-arrow-black.svg";
+import { ReactComponent as BlueRightArrow } from "../assets/right-arrow-blue.svg";
+import * as S from "../styles";
+
+function GivingsSection(): JSX.Element {
+  const [allDonations, setAllDonations] = useState<any>();
+  const { navigateTo } = useNavigation();
+  const { t } = useTranslation("translation", {
+    keyPrefix: "promoters.fundPage.givingsSection",
+  });
+  const { wallet, connectWallet } = useWalletContext();
+  const { getAllPromotersDonations } = usePromoterDonations();
+  const { isMobile } = useBreakpoint();
+  const { currentNetwork } = useNetwork();
+  const coin = "USDC";
+
+  const handleShowGivingsButtonClick = () => {
+    logEvent("fundShowGivingsListBtn_click", {
+      from: "yourGivingsCarousel",
+    });
+    if (wallet) {
+      navigateTo("/promoters/show-givings");
+      return;
+    }
+    connectWallet();
+  };
+
+  const fetchAllDonations = useCallback(async () => {
+    try {
+      const donations = await getAllPromotersDonations(isMobile ? 2 : 3);
+      setAllDonations(donations.promoterDonations);
+    } catch (e) {
+      logError(e);
+    }
+  }, []);
+
+  useEffect(() => {
+    fetchAllDonations();
+  }, []);
+
+  function concatLinkHash(hash: string) {
+    return `${currentNetwork.blockExplorerUrls}tx/${hash}`;
+  }
+
+  function renderCardsCarouselAllGivings() {
+    return allDonations?.map((item: any) => (
+      <CardDoubleTextDividerButton
+        key={item.id}
+        firstText={formatDate(item.timestamp).toString()}
+        mainText={formatFromWei(item.amountDonated)}
+        rightComplementText={coin}
+        buttonText={t("linkTransactionText")}
+        rightComponentButton={RightArrowBlack}
+        link={concatLinkHash(item.id)}
+      />
+    ));
+  }
+
+  return (
+    <S.Container>
+      <S.SectionTitle>All Givings</S.SectionTitle>
+      <Carousel sliderPerView={isMobile ? 1.8 : 4} spacing={-10}>
+        {renderCardsCarouselAllGivings()}
+        {false && (
+          <S.LastCardCarousel
+            onClick={() => {
+              handleShowGivingsButtonClick();
+            }}
+          >
+            <BlueRightArrow />
+            <S.TextLastCard>{t("textLastCard")}</S.TextLastCard>
+          </S.LastCardCarousel>
+        )}
+      </Carousel>
+    </S.Container>
+  );
+}
+export default GivingsSection;
